# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DOCUMENTATION",
        "description": "Bookings and Countries API",
        "contact": {
            "email": "nissa.khairunnisaaa@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get list of bookings",
                "description": "Returns a list of bookings",
                "operationId": "getBookingsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "description": "Stores a new booking along with associated member, id verification, and country records",
                "operationId": "storeBooking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "member_id": {
                                        "description": "ID of the member making the booking",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "country_id": {
                                        "description": "ID of the country for the booking",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "surfing_experience": {
                                        "description": "Surfing experience level (1-10+)",
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "visit_date": {
                                        "description": "Date of the booking",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-10"
                                    },
                                    "desired_board": {
                                        "description": "Desired type of surfboard",
                                        "type": "string",
                                        "enum": [
                                            "longboard",
                                            "funboard",
                                            "shortboard",
                                            "fishboard",
                                            "gunboard"
                                        ],
                                        "example": "longboard"
                                    },
                                    "id_card_image": {
                                        "description": "Image file of the ID card for verification",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "link_url_path": {
                                        "description": "link image url of ID verification",
                                        "type": "string",
                                        "example": "http://example.com/1.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Booking"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "member_id": [
                                                    "The member id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings/{id}": {
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update an existing booking",
                "description": "Updates an existing booking",
                "operationId": "updateBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "member_id": {
                                        "description": "ID of the member making the booking",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "country_id": {
                                        "description": "ID of the country for the booking",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "surfing_experience": {
                                        "description": "Surfing experience level (1-10+)",
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "visit_date": {
                                        "description": "Date of the booking",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-10"
                                    },
                                    "desired_board": {
                                        "description": "Desired type of surfboard",
                                        "type": "string",
                                        "enum": [
                                            "longboard",
                                            "funboard",
                                            "shortboard",
                                            "fishboard",
                                            "gunboard"
                                        ],
                                        "example": "longboard"
                                    },
                                    "id_card_image": {
                                        "description": "Image file of the ID card for verification",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "link_url_path": {
                                        "description": "link image url of ID verification",
                                        "type": "string",
                                        "example": "http://example.com/1.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a booking",
                "description": "Deletes a booking",
                "operationId": "deleteBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking deleted successfully"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Booking": {},
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}